adds check to cronjob

+++ src/main/java/io/supertokens/cronjobs/deleteExpiredSessions/DeleteExpiredSessions.java
@@ -20,6 +20,7 @@ import io.supertokens.Main;
 import io.supertokens.ResourceDistributor;
 import io.supertokens.cronjobs.CronTask;
 import io.supertokens.cronjobs.CronTaskTest;
+import io.supertokens.pluginInterface.STORAGE_TYPE;
 import io.supertokens.storageLayer.StorageLayer;
 
 public class DeleteExpiredSessions extends CronTask {
@@ -40,6 +41,9 @@ public class DeleteExpiredSessions extends CronTask {
 
     @Override
     protected void doTask() throws Exception {
+        if (StorageLayer.getStorage(this.main).getType() != STORAGE_TYPE.SQL) {
+            return;
+        }
         StorageLayer.getSessionStorage(this.main).deleteAllExpiredSessions();
     }
